Adds a shuffle and play button for albums
Implements a new "Shuffle & Play" button on the album view for mobile devices.

This button allows users to immediately start playing a random selection of songs from their library, improving the user experience by providing a quick way to start listening to music.

The button is only visible on mobile devices to keep the UI clean on larger screens, and a similar button has been added to the AppBar for larger screens.