Adds mobile player UI component
Introduces a mobile-specific player bar that appears at the bottom of the screen on mobile devices. This player provides basic playback controls and track information.

This change also introduces a setting to allow the user to switch between mini and full player modes, with the mini player enabled by default.

The full player is now always rendered to allow for audio playback, but is hidden or shown based on selected player mode.